+++ hw09_struct_validator/consts.go
@@ -1,0 +1,13 @@
+package hw09structvalidator
+
+const (
+	validateTag        = "validate"
+	separator          = "|"
+	valueSeparator     = ","
+	tagSeparator       = ":"
+	lenTagValidator    = "len"
+	inTagValidator     = "in"
+	regexpTagValidator = "regexp"
+	minTagValidator    = "min"
+	maxTagValidator    = "max"
+)
